flask/requirements.txt
aniso8601==8.0.0
attrs==19.1.0
awsebcli==3.15.3
blessed==1.15.0
blis==0.2.4
botocore==1.12.234
cached-property==1.5.1
cachetools==3.1.0
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
cymem==2.0.2
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
ez-setup==0.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
future==0.16.0
graphviz==0.11
gunicorn==19.9.0
httplib2==0.12.3
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
pathspec==0.5.9
patsy==0.5.1
plac==0.9.6
postgres==2.2.2
preshed==2.0.1
protobuf==3.8.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
py-openaq==1.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.3.1
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.3.10
squarify==0.4.3
srsly==0.0.7
statsmodels==0.10.1
tabulate==0.8.3
termcolor==1.1.0
texttable==0.9.1
thinc==7.0.8
tqdm==4.36.1
urllib3==1.24.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
wasabi==0.2.2
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.5
